Add support for Thunar file manager

+++ install.sh
@@ -105,6 +105,10 @@ _check_default_filemanager() {
         INSTALL_DIR="$HOME/.config/caja/scripts"
         ACCELS_FILE="$HOME/.config/caja/accels"
         FILE_MANAGER="caja"
+    elif _command_exists "thunar"; then
+        ACCELS_FILE="$HOME/.config/Thunar/accels.scm"
+        FILE_MANAGER="thunar"
+        INSTALL_DIR="$HOME/.local/scripts"
     else
         printf "Error: could not find any compatible file managers!\n"
         exit 1
@@ -142,15 +146,13 @@ _step_install_dependencies() {
         if _command_exists "apt-get"; then
             # Distro: Ubuntu, Mint, Debian.
             sudo apt-get update || true
-            sudo apt-get -y install $common_names p7zip-full imagemagick xz-utils poppler-utils ffmpeg findimagedupes genisoimage
+            sudo apt-get -y install $common_names p7zip-full imagemagick xz-utils poppler-utils ffmpeg genisoimage
         elif _command_exists "dnf"; then
             # Distro: Fedora, Red Hat.
-            # Missing packages: findimagedupes.
             sudo dnf check-update || true
             sudo dnf -y install $common_names p7zip ImageMagick xz poppler-utils ffmpeg-free genisoimage
         elif _command_exists "pacman"; then
             # Distro: Manjaro, Arch Linux.
-            # Missing packages: findimagedupes.
             sudo pacman -Syy || true
             sudo pacman --noconfirm -S $common_names p7zip imagemagick xz poppler poppler-glib ffmpeg
         else
@@ -213,13 +215,23 @@ _step_install_scripts() {
         printf " > Restoring previous scripts to the install directory...\n"
         mv "$tmp_install_dir/scripts" "$INSTALL_DIR/User previous scripts"
     fi
+
+    # Install the menus for the 'Thunar' file manager.
+    if [[ "$FILE_MANAGER" == "thunar" ]]; then
+        printf " > Installing Thunar custom actions...\n"
+        _step_make_thunar_custom_actions
+    fi
 }
 
 _step_install_shortcuts() {
     printf " > Installing the keyboard shortcuts...\n"
 
     mkdir --parents "$(dirname -- "$ACCELS_FILE")"
-    mv "$ACCELS_FILE" "$ACCELS_FILE.bak" 2>/dev/null || true
+
+    # Create a backup of older shortcuts.
+    if [[ -f "$ACCELS_FILE" ]] && ! [[ -f "$ACCELS_FILE.bak" ]]; then
+        mv "$ACCELS_FILE" "$ACCELS_FILE.bak" 2>/dev/null || true
+    fi
 
     case "$FILE_MANAGER" in
     "nautilus")
@@ -227,13 +239,18 @@ _step_install_shortcuts() {
         ;;
     "nemo")
         cp -- "$ASSSETS_DIR/accels-mint" "$ACCELS_FILE"
-        sed -i "s|USER|$USER|g" "$ACCELS_FILE"
-        sed -i "s|ACCELS_PATH|local\\\\\\\\sshare\\\\\\\\snemo|g" "$ACCELS_FILE"
+        sed -i "s|SED_USER|$USER|g" "$ACCELS_FILE"
+        sed -i "s|SED_ACCELS_PATH|local\\\\\\\\sshare\\\\\\\\snemo|g" "$ACCELS_FILE"
         ;;
     "caja")
         cp -- "$ASSSETS_DIR/accels-mint" "$ACCELS_FILE"
-        sed -i "s|USER|$USER|g" "$ACCELS_FILE"
-        sed -i "s|ACCELS_PATH|config\\\\\\\\scaja|g" "$ACCELS_FILE"
+        sed -i "s|SED_USER|$USER|g" "$ACCELS_FILE"
+        sed -i "s|SED_ACCELS_PATH|config\\\\\\\\scaja|g" "$ACCELS_FILE"
+        ;;
+    "thunar")
+        cp -- "$ASSSETS_DIR/accels-thunar" "$ACCELS_FILE"
+        sed -i "s|SED_USER|$USER|g" "$ACCELS_FILE"
+        sed -i "s|SED_ACCELS_PATH|config\\\\\\\\scaja|g" "$ACCELS_FILE"
         ;;
     esac
 }
@@ -242,6 +259,84 @@ _step_close_filemanager() {
     printf " > Closing the file manager to reload its configurations...\n"
 
     eval "$FILE_MANAGER -q &>/dev/null" || true
+}
+
+_step_make_thunar_custom_actions() {
+    local menus_file="$HOME/.config/Thunar/uca.xml"
+
+    # Create a backup of older custom actions.
+    if [[ -f "$menus_file" ]] && ! [[ -f "$menus_file.bak" ]]; then
+        mv "$menus_file" "$menus_file.bak" 2>/dev/null || true
+    fi
+
+    {
+        printf "%s\n" "<?xml version="1.0" encoding="UTF-8"?>"
+        printf "%s\n" "<actions>"
+        printf "%s\n" "<action>"
+        printf "%s\n" "    <icon>utilities-terminal</icon>"
+        printf "%s\n" "    <name>Open Terminal Here</name>"
+        printf "%s\n" "    <submenu></submenu>"
+        printf "%s\n" "    <unique-id>1-1</unique-id>"
+        printf "%s\n" "    <command>exo-open --working-directory %f --launch TerminalEmulator</command>"
+        printf "%s\n" "    <description>Open terminal in containing directory</description>"
+        printf "%s\n" "    <range></range>"
+        printf "%s\n" "    <patterns>*</patterns>"
+        printf "%s\n" "    <startup-notify/>"
+        printf "%s\n" "    <directories/>"
+        printf "%s\n" "</action>"
+        printf "%s\n" "<action>"
+        printf "%s\n" "    <icon>edit-find</icon>"
+        printf "%s\n" "    <name>Find in this folder</name>"
+        printf "%s\n" "    <submenu></submenu>"
+        printf "%s\n" "    <unique-id>3-3</unique-id>"
+        printf "%s\n" "    <command>catfish --path=%f</command>"
+        printf "%s\n" "    <description>Search for files within this folder</description>"
+        printf "%s\n" "    <range></range>"
+        printf "%s\n" "    <patterns>*</patterns>"
+        printf "%s\n" "    <directories/>"
+        printf "%s\n" "</action>"
+        printf "%s\n" "<action>"
+        printf "%s\n" "    <icon>document-print</icon>"
+        printf "%s\n" "    <name>Print file(s)</name>"
+        printf "%s\n" "    <submenu></submenu>"
+        printf "%s\n" "    <unique-id>4-4</unique-id>"
+        printf "%s\n" "    <command>thunar-print %F</command>"
+        printf "%s\n" "    <description>Send one or multiple files to the default printer</description>"
+        printf "%s\n" "    <range></range>"
+        printf "%s\n" "    <patterns>*.asc;*.brf;*.css;*.doc;*.docm;*.docx;*.dotm;*.dotx;*.fodg;*.fodp;*.fods;*.fodt;*.gif;*.htm;*.html;*.jpe;*.jpeg;*.jpg;*.odb;*.odf;*.odg;*.odm;*.odp;*.ods;*.odt;*.otg;*.oth;*.otp;*.ots;*.ott;*.pbm;*.pdf;*.pgm;*.png;*.pnm;*.pot;*.potm;*.potx;*.ppm;*.ppt;*.pptm;*.pptx;*.rtf;*.shtml;*.srt;*.text;*.tif;*.tiff;*.txt;*.xbm;*.xls;*.xlsb;*.xlsm;*.xlsx;*.xltm;*.xltx;*.xpm;*.xwd</patterns>"
+        printf "%s\n" "    <image-files/>"
+        printf "%s\n" "    <other-files/>"
+        printf "%s\n" "    <text-files/>"
+        printf "%s\n" "</action>"
+
+        local filename=""
+        local name=""
+        local submenu=""
+        local unique_id=""
+        find "$INSTALL_DIR" -mindepth 2 -type f ! -path "*.git/*" ! -path "*.assets/*" -print0 2>/dev/null | sort --zero-terminated |
+            while IFS= read -r -d "" filename; do
+                name=$(basename -- "$filename" 2>/dev/null)
+                submenu=$(dirname -- "$filename" 2>/dev/null | sed "s|.*scripts/|Scripts/|g")
+                unique_id=$(md5sum <<<"$name" 2>/dev/null | tr -d "a-z- ")
+                printf "%s\n" "<action>"
+                printf "%s\n" "    <icon></icon>"
+                printf "%s\n" "    <name>$name</name>"
+                printf "%s\n" "    <submenu>$submenu</submenu>"
+                printf "%s\n" "    <unique-id>$unique_id-1</unique-id>"
+                printf "%s\n" "    <command>bash &quot;$filename&quot; %F</command>"
+                printf "%s\n" "    <description></description>"
+                printf "%s\n" "    <patterns>*</patterns>"
+                printf "%s\n" "    <directories/>"
+                printf "%s\n" "    <audio-files/>"
+                printf "%s\n" "    <image-files/>"
+                printf "%s\n" "    <other-files/>"
+                printf "%s\n" "    <text-files/>"
+                printf "%s\n" "    <video-files/>"
+                printf "%s\n" "</action>"
+            done
+
+        printf "%s\n" "<actions>"
+    } >"$menus_file"
 }
 
 _main "$@"
